streamlit-app/requirements.txt
streamlit>=1.28.0
pandas>=2.0.3
numpy>=1.24.3
faiss-cpu>=1.7.2
sentence-transformers>=2.2.2
torch>=2.0.1
Pillow>=10.0.0
huggingface-hub>=0.19.0
